added releaseName template and .Values.global.releaseNameOverride variable
Replaced all occurrences of .Release.Name with the template invocation. The releaseNameOverride variable allows to override a helm release name that supposed to be a free  and nonconflicting. This allows to easily add and delete new deployments parts to a already existing deployment. Allows for quicker more iterative development and testing.